fix shared dependencies that are not resident in a database
eg. databases depend on users (their owners) that both don’t have a
database they reside in. These dependencies are recorded in pg_shdepend
with a dbid of 0. When we fetch our shared dependencies we don’t take
these links in account.

With this patch we use logic inspired by classIdGetDbId to decide when
to use MyDatabaseId vs InvalidOid to correctly resolve dependencies
between shared objects.