Improve regression tests for prepared statements
With a recent commit, we made (https://github.com/citusdata/citus/commit/644b266deed2e9dc32664ffe75c902424d2f3361)
the behaviour of prepared statements for local cached plans has
slightly changed.

Now, Citus caches the plans when they are re-used. This make triggering
of local cached plans on the 7th execution, and 8th execution is the
first time the plan is used from the cached.

So, the tests are improved to cover 8th execution.