Fix distributed deadlock with partitions
If we have a multi shard update on a partition and a concurrent
drop/create partition a distributed deadlock was possible.

The reason for the deadlock is Postgres takes an AccessExclusiveLock on
the parent table when dropping a partition, however it might not take
AccessShareLock on the parent table when updating a partition if the
child quals are already cached.

When AccessShareLock is not taken on the parent, Drop command will be
sent to workers and depending on the timing a DROP command can take an
AccessExclusiveLock on one worker and an UPDATE can take AccessShareLock
on another worker blocking each other to process.

To resolve the issue, we now take an AccessExclusiveLock on the parent table in case of
modify commands. This makes sure that DROP commands will wait.