Fix crash in shard rebalancer when no distributed tables exist
The logging of the amount of ignored moves crashed when a no distributed
tables existed in a cluster. This also fixes in passing that the logging
of ignored moves logs the correct number of ignored moves if there
exist multiple colocation groups and all are rebalanced at the same time.