(Share) Lock buffer page when reading from columnar storage
Under high write concurrency, we were sometimes reading columnar
metapage as all zeros.

We still couldn't find the exact reason, however
postgres/storage/buffer/README states that:
```
Buffer access rules:

1. To scan a page for tuples, one must hold a pin and either shared or
exclusive content lock.  To examine the commit status (XIDs and status
bits)
of a tuple in a shared buffer, one must likewise hold a pin and either
shared
or exclusive lock.
```

That said, it seems quite possible that a transaction that needs to read
metapage might see dirty values if there is a concurrent writer
transaction that needs to update reserved.* fields of metapage.

Plus, when testing this commit, we didn't encounter with all-zero
metapages anymore.

However, it is still strange that we couldn't reproduce this issue on
older versions even if we don't (Share) lock buffer page when reading
columnar tables on those versions too.

On the other hand, with the changes we made around columnar storage
layer and read-path on 10.2, now we are inclined to fetch metapage more
frequently, so this could be the reason why this issue only applies to
10.2.