Add missing downgrade path from 10.2-4 to 11.x
We recently introduced 10.2-4 in a patch release for 10.2. This is not a
migration version that exists on our default branch, and hence we do not
have a migration path from 10.2-4 to the current latest 11.0-1.

I intentionally did not place the downgrade script inside the
`downgrades/` folder because we need this migration script to be
installed by `make install`.