Sync metadata in new clusters
If a user upgrades from an ealier version, and have not synced the
metadata, we should not automatically sync the metadata.

If this is a post Citus 11 cluster, and we are adding a new worker node,
we sync the metadata.

The net effect is that new clusters gets Citus MX by default, existing
clusters should manually trigger metadata syncing.