Implement TEXT SEARCH CONFIGURATION propagation
The change adds support to Citus for propagating TEXT SEARCH CONFIGURATION objects.
TSConfig objects cannot always be created in one create statement, and instead
require a create statement followed by many alter statements to get turned into
the object they should represent.

To support this we add functionality to the worker to create or replace objects
based on a list of statements. When the lists of the local object and the remote
object correspond 1:1 we skip the creation of the object and simply mark it
distributed. This is especially important for TSConfig objects as initdb pre-
populates databases with a dozen configurations (for many different languages).

When the user creates a new TSConfig based on the copy of an existing configuration
there is no direct link to the object copied from. Since there is no link we can't
simply rely on propagating the dependencies to the worker and send a qualified
statement over. Instead we generate the full list of create + alter statments
based on what we find in our own catalog after the object is created. eg,
creating based on a copy of the `dutch` configuration will result in ALTER
statements for every one of the mappings that configuration had.