Copy errmsg for distributed deadlock error into heap
multi_log_hook hook called by EmitErrorReport() when emitting the ereport
either to frontend or to the server logs. And some callers of
EmitErrorReport() (e.g.:errfinish) seems to assume that string fields of
given ErrorData object needs to be freed. For this reason, we copy the
message into heap here.